refactor bank staff business logic to service module

+++ bank_staff/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+from decimal import Decimal
+
 # Create your models here.
 
 
@@ -7,13 +9,13 @@ class Bank(models.Model):
     total_funds = models.DecimalField(max_digits=19, decimal_places=2)
 
     @classmethod
-    def add_funds(cls, amount: float):
+    def add_funds(cls, amount: Decimal):
         bank = cls.objects.get(pk=1)
         bank.total_funds += amount
         bank.save()
 
     @classmethod
-    def remove_funds(cls, amount: float):
+    def remove_funds(cls, amount: Decimal):
         bank = cls.objects.get(pk=1)
         bank.total_funds -= amount
         bank.save()
